Move hash functions to their own private module
There are a few of these functions so they should be in their own
module. The API is unchanged, this is just internal organization. Took
this chance to unify and better parametrize the hash tests.
